move cp into osutil (thanks Gustavo)
