+++ coreconfig/config.go
@@ -32,8 +32,8 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/ubuntu-core/snappy/cp"
 	"github.com/ubuntu-core/snappy/helpers"
+	"github.com/ubuntu-core/snappy/osutil"
 	"github.com/ubuntu-core/snappy/systemd"
 
 	"gopkg.in/yaml.v2"
@@ -319,7 +319,7 @@ var getTimezone = func() (timezone string, err error) {
 // setTimezone sets the specified timezone for the system, an error is returned
 // if it can't.
 var setTimezone = func(timezone string) error {
-	if err := cp.CopyFile(filepath.Join(tzZoneInfoPath, timezone), tzZoneInfoTarget, cp.CopyFlagOverwrite); err != nil {
+	if err := osutil.CopyFile(filepath.Join(tzZoneInfoPath, timezone), tzZoneInfoTarget, osutil.CopyFlagOverwrite); err != nil {
 		return err
 	}
