+++ osutil/cp_linux_test.go
@@ -1,5 +1,4 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
-// +build !linux
 
 /*
  * Copyright (C) 2014-2015 Canonical Ltd
@@ -18,14 +17,30 @@
  *
  */
 
-package cp
+package osutil
 
 import (
-	"io"
+	"io/ioutil"
 	"os"
+
+	. "gopkg.in/check.v1"
 )
 
-func doCopyFile(fin, fout fileish, fi os.FileInfo) error {
-	_, err := io.Copy(fout, fin)
-	return err
+func (s *cpSuite) TestCpMulti(c *C) {
+	maxcp = 2
+	defer func() { maxcp = maxint }()
+
+	c.Check(CopyFile(s.f1, s.f2, CopyFlagDefault), IsNil)
+	bs, err := ioutil.ReadFile(s.f2)
+	c.Check(err, IsNil)
+	c.Check(bs, DeepEquals, s.data)
+}
+
+func (s *cpSuite) TestDoCpErr(c *C) {
+	f1, err := os.Open(s.f1)
+	c.Assert(err, IsNil)
+	st, err := f1.Stat()
+	c.Assert(err, IsNil)
+	// force an error by asking it to write to a readonly stream
+	c.Check(doCopyFile(f1, os.Stdin, st), NotNil)
 }
