+++ osutil/cp_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package cp
+package osutil
 
 import (
 	"errors"
@@ -27,13 +27,10 @@ import (
 	"path/filepath"
 	"strings"
 	"syscall"
-	"testing"
 	"time"
 
 	. "gopkg.in/check.v1"
 )
-
-func Test(t *testing.T) { TestingT(t) }
 
 type cpSuite struct {
 	dir  string
